[#1152] Refactor backup file utilities into dedicated modules
