+++ src/include/info.h
@@ -89,31 +89,6 @@ extern "C" {
 
 #define INFO_BUFFER_SIZE               8192
 
-/**
- * @struct rfile
- * An rfile stores the metadata we need to use a file on disk for reconstruction.
- * For full backup file in the chain, only file name and file pointer are initialized.
- *
- * extracted flag indicates if the file is a copy extracted from the original file
- * num_blocks is the number of blocks present inside an incremental file.
- * These are the blocks that have changed since the last checkpoint.
- * truncation_block_length is basically the shortest length this file has been between this and last checkpoint.
- * Note that truncation_block_length could be even greater than the number of blocks the original file has.
- * Because the tables are not locked during the backup, so blocks could be truncated during the process,
- * while truncation_block_length only reflects length until the checkpoint before backup starts.
- * relative_block_numbers are the relative BlockNumber of each block in the file. Relative here means relative to
- * the starting BlockNumber of this file.
- */
-struct rfile
-{
-   char* filepath;                   /**< The path of the backup file  */
-   FILE* fp;                         /**< The file descriptor corresponding to the backup file */
-   size_t header_length;             /**< The header length */
-   uint32_t num_blocks;              /**< The number of blocks present inside an incremental file */
-   uint32_t* relative_block_numbers; /**< relative_block_numbers are the relative BlockNumber of each block in the file */
-   uint32_t truncation_block_length; /**< truncation_block_length only reflects length until the checkpoint before backup starts. */
-};
-
 /** @struct backup
  * Defines a backup
  */
@@ -263,41 +238,6 @@ pgmoneta_annotate_request(SSL* ssl, int client_fd, int server, uint8_t compressi
  */
 int
 pgmoneta_save_info(char* directory, struct backup* backup);
-
-/**
- * Create an rfile structure of a backup file
- * @param server The server
- * @param label The label of the backup
- * @param relative_dir The relative path inside the data directory (excluding the filename)
- * @param base_file_name The file name
- * @param encryption The encryption method
- * @param compression The compression method
- * @param rfile [out] The rfile
- * @return 0 if success, otherwise 1
- */
-int
-pgmoneta_rfile_create(int server, char* label, char* relative_dir, char* base_file_name, int encryption, int compression, struct rfile** rfile);
-
-/**
- * Destroy the rfile structure
- * @param rfile The rfile to be destroyed
- */
-void
-pgmoneta_rfile_destroy(struct rfile* rf);
-
-/**
- * Initialize an rfile structure of an incremental file by reading the incremental file headers
- * @param server The server
- * @param label The label of the backup
- * @param relative_dir The relative path inside the data directory (excluding the filename)
- * @param base_file_name The file name
- * @param encryption The encryption method
- * @param compression The compression method
- * @param rfile [out] The rfile
- * @return 0 if success, otherwise 1
- */
-int
-pgmoneta_incremental_rfile_initialize(int server, char* label, char* relative_dir, char* base_file_name, int encryption, int compression, struct rfile** rfile);
 
 /**
  * Get an approximate size of a backup repository
