+++ src/include/utils.h
@@ -35,7 +35,7 @@ extern "C" {
 
 #include <pgmoneta.h>
 #include <art.h>
-#include <extraction.h>
+#include <files.h>
 #include <message.h>
 #include <workers.h>
