+++ src/libpgmoneta/info.c
@@ -27,22 +27,22 @@
  */
 
 /* pgmoneta */
-#include <assert.h>
 #include <pgmoneta.h>
 #include <backup.h>
-#include <extraction.h>
 #include <info.h>
 #include <logging.h>
 #include <management.h>
 #include <network.h>
-#include <stddef.h>
-#include <stdint.h>
-#include <utils.h>
+#include <rfile.h>
 #include <security.h>
+#include <utils.h>
 
 /* system */
+#include <assert.h>
 #include <errno.h>
 #include <libgen.h>
+#include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -58,9 +58,6 @@
  */
 static void __attribute__((unused))
 create_info(char* directory, char* label, int status);
-
-static int
-file_final_name(char* file, int encryption, int compression, char** finalname);
 
 /**
  * Best effort to split a file path into a relative path and a bare file name
@@ -1456,179 +1453,6 @@ error:
 }
 
 int
-pgmoneta_rfile_create(int server, char* label, char* relative_dir, char* base_file_name, int encryption, int compression, struct rfile** rfile)
-{
-   struct rfile* rf = NULL;
-   char* extracted_file_path = NULL;
-   char* final_relative_path = NULL;
-   char base_relative_path[MAX_PATH];
-   FILE* fp = NULL;
-
-   memset(base_relative_path, 0, MAX_PATH);
-   if (pgmoneta_ends_with(relative_dir, "/"))
-   {
-      pgmoneta_snprintf(base_relative_path, MAX_PATH, "%s%s", relative_dir, base_file_name);
-   }
-   else
-   {
-      pgmoneta_snprintf(base_relative_path, MAX_PATH, "%s/%s", relative_dir, base_file_name);
-   }
-
-   /* try both base and final relative path */
-   if (pgmoneta_extract_backup_file(server, label, base_relative_path, NULL, &extracted_file_path))
-   {
-      free(extracted_file_path);
-      extracted_file_path = NULL;
-      file_final_name(base_relative_path, encryption, compression, &final_relative_path);
-      if (pgmoneta_extract_backup_file(server, label, final_relative_path, NULL, &extracted_file_path))
-      {
-         goto error;
-      }
-   }
-   fp = fopen(extracted_file_path, "r");
-
-   if (fp == NULL)
-   {
-      goto error;
-   }
-   rf = (struct rfile*)malloc(sizeof(struct rfile));
-   memset(rf, 0, sizeof(struct rfile));
-
-   rf->fp = fp;
-   rf->filepath = extracted_file_path;
-   *rfile = rf;
-
-   free(final_relative_path);
-   return 0;
-
-error:
-   free(extracted_file_path);
-   free(final_relative_path);
-   pgmoneta_rfile_destroy(rf);
-   return 1;
-}
-
-void
-pgmoneta_rfile_destroy(struct rfile* rf)
-{
-   if (rf == NULL)
-   {
-      return;
-   }
-   if (rf->fp != NULL)
-   {
-      fclose(rf->fp);
-   }
-   if (rf->filepath != NULL)
-   {
-      // this is the extracted file, we should delete it
-      pgmoneta_delete_file(rf->filepath, NULL);
-   }
-
-   free(rf->filepath);
-   free(rf->relative_block_numbers);
-   free(rf);
-}
-
-int
-pgmoneta_incremental_rfile_initialize(int server, char* label, char* relative_dir, char* base_file_name, int encryption, int compression, struct rfile** rfile)
-{
-   uint32_t magic = 0;
-   uint32_t nread = 0;
-   struct rfile* rf = NULL;
-   struct main_configuration* config;
-   size_t relsegsz = 0;
-   size_t blocksz = 0;
-
-   config = (struct main_configuration*)shmem;
-
-   relsegsz = config->common.servers[server].relseg_size;
-   blocksz = config->common.servers[server].block_size;
-
-   /*
-    * Header structure:
-    * magic number(uint32)
-    * num blocks (number of changed blocks, uint32)
-    * truncation block length (uint32)
-    * relative_block_numbers (uint32 * (num blocks))
-    */
-
-   // create rfile after file is opened successfully
-   if (pgmoneta_rfile_create(server, label, relative_dir, base_file_name, encryption, compression, &rf))
-   {
-      pgmoneta_log_error("rfile initialize: failed to open incremental backup (label %s) file at %s/%s", label, relative_dir, base_file_name);
-      goto error;
-   }
-
-   // read magic number from header
-   nread = fread(&magic, 1, sizeof(uint32_t), rf->fp);
-   if (nread != sizeof(uint32_t))
-   {
-      pgmoneta_log_error("rfile initialize: incomplete file header at %s, cannot read magic number", rf->filepath);
-      goto error;
-   }
-
-   if (magic != INCREMENTAL_MAGIC)
-   {
-      pgmoneta_log_error("rfile initialize: incorrect magic number, getting %X, expecting %X", magic, INCREMENTAL_MAGIC);
-      goto error;
-   }
-
-   // read number of blocks
-   nread = fread(&rf->num_blocks, 1, sizeof(uint32_t), rf->fp);
-   if (nread != sizeof(uint32_t))
-   {
-      pgmoneta_log_error("rfile initialize: incomplete file header at %s%s, cannot read block count", relative_dir, base_file_name);
-      goto error;
-   }
-   if (rf->num_blocks > relsegsz)
-   {
-      pgmoneta_log_error("rfile initialize: file has %d blocks which is more than server's segment size", rf->num_blocks);
-      goto error;
-   }
-
-   // read truncation block length
-   nread = fread(&rf->truncation_block_length, 1, sizeof(uint32_t), rf->fp);
-   if (nread != sizeof(uint32_t))
-   {
-      pgmoneta_log_error("rfile initialize: incomplete file header at %s%s, cannot read truncation block length", relative_dir, base_file_name);
-      goto error;
-   }
-   if (rf->truncation_block_length > relsegsz)
-   {
-      pgmoneta_log_error("rfile initialize: file has truncation block length of %d which is more than server's segment size", rf->truncation_block_length);
-      goto error;
-   }
-
-   if (rf->num_blocks > 0)
-   {
-      rf->relative_block_numbers = malloc(sizeof(uint32_t) * rf->num_blocks);
-      nread = fread(rf->relative_block_numbers, sizeof(uint32_t), rf->num_blocks, rf->fp);
-      if (nread != rf->num_blocks)
-      {
-         pgmoneta_log_error("rfile initialize: incomplete file header at %s, cannot read relative block numbers", rf->filepath);
-         goto error;
-      }
-   }
-
-   // magic + block num + truncation block length + relative block numbers
-   rf->header_length = sizeof(uint32_t) * (1 + 1 + 1 + rf->num_blocks);
-   // round header length to multiple of block size, since the actual file data are aligned
-   // only needed when the file actually has data
-   if (rf->num_blocks > 0 && rf->header_length % blocksz != 0)
-   {
-      rf->header_length += (blocksz - (rf->header_length % blocksz));
-   }
-
-   *rfile = rf;
-   return 0;
-error:
-   // contains fp closing logic
-   pgmoneta_rfile_destroy(rf);
-   return 1;
-}
-
-int
 pgmoneta_backup_size(int server, char* label, unsigned long* size, uint64_t* biggest_file_size)
 {
    struct json* manifest_read = NULL;
@@ -1766,40 +1590,6 @@ pgmoneta_sort_backups(struct backup** backups, int number_of_backups, bool desc)
    return 0;
 
 error:
-   return 1;
-}
-
-static int
-file_final_name(char* file, int encryption, int compression, char** finalname)
-{
-   char* final = NULL;
-
-   *finalname = NULL;
-   if (file == NULL)
-   {
-      goto error;
-   }
-
-   final = pgmoneta_append(final, file);
-   {
-      char* suffix = NULL;
-
-      if (pgmoneta_extraction_get_suffix(compression, encryption, &suffix))
-      {
-         goto error;
-      }
-      if (suffix != NULL)
-      {
-         final = pgmoneta_append(final, suffix);
-      }
-      free(suffix);
-   }
-
-   *finalname = final;
-   return 0;
-
-error:
-   free(final);
    return 1;
 }
